Add files via upload
This version includes a varient AEA0 which is the Blade Area Ratio of the Bseries Prop. It shows that the design tends to look like a modern windturbine at the higher efficiency points. as high radius as possible with minimal blade area; large, long and thin and slowmoving blades. an single optimal Bseries prop should end up around an openwater efficiency of 0.6 in a Realistic Marine application.